🎨 style(profile.png): add profile image to main page The profile image has been added to the main page of the website.
🗑️ chore(profile.dart): delete profile.dart file
The profile.dart file has been deleted and is no longer needed in the project.

🎨 style(home_view.dart): rename profile.dart to home_profile.dart
🚀 feat(home_profile.dart): add Avatar widget
The profile.dart file was renamed to home_profile.dart to better reflect its purpose. The Avatar widget was added to display the user's profile image. The widget uses the AvatarGlow package to add a glowing effect to the image. The CachedNetworkImage package is used to load the image from the network and display a placeholder while the image is loading.

✨ feat(user_input.dart): add UserInput widget to Home Profile
This commit adds the UserInput widget to the Home Profile screen. The widget allows the user to view and edit their display name and email address. It also displays the user's unique ID and whether their email address has been verified. The widget uses the FormBuilderTextField widget to create the input fields and the AwesomeDialog package to display success and error messages. The widget is controlled by the HomeController and uses the FirebaseAuth package to authenticate the user.

🎉 feat(home_profile.dart): add FormFirebase widget
🚀 chore(pubspec.yaml): add avatar_glow package
The FormFirebase widget is added to the home_profile.dart file. This widget contains a form that allows the user to input their details and link their Google account. The avatar_glow package is added to the pubspec.yaml file to support the Avatar widget used in the FormFirebase widget.
