+++ pubspec.yaml
@@ -7,9 +7,10 @@ environment:
 
 dependencies:
   about: ^2.1.1
+  avatar_glow: ^2.0.2
   awesome_dialog: ^3.0.2
-  bottom_bar_matu: ^1.3.0
   bootstrap_alert: ^1.0.1
+  bottom_bar_matu: ^1.3.0
   cached_network_image: ^3.2.3
   collection: 1.17.1
   cupertino_icons: ^1.0.5
